+++ lib/src/emoji_picker_internal_utils.dart
@@ -80,6 +80,12 @@ class EmojiPickerInternalUtils {
     return recentEmoji;
   }
 
+  /// Clears the list of recent emojis in local storage
+  Future<void> clearRecentEmojisInLocalStorage() async {
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setString('recent', jsonEncode([]));
+  }
+
   /// Remove skin tone from given emoji
   Emoji removeSkinTone(Emoji emoji) {
     return emoji.copyWith(