+++ lib/src/emoji_picker_utils.dart
@@ -97,4 +97,12 @@ class EmojiPickerUtils {
     }
     return emoji.copyWith(emoji: String.fromCharCodes(result));
   }
+
+  /// Clears the list of recent emojis
+  Future<void> clearRecentEmojis(
+      {required GlobalKey<EmojiPickerState> key}) async {
+    return await EmojiPickerInternalUtils()
+        .clearRecentEmojisInLocalStorage()
+        .then((_) => key.currentState?.updateRecentEmoji([], refresh: true));
+  }
 }